+++ packages/sources/src/components/form/content/input-number.content.ts
@@ -1,0 +1,14 @@
+import { FormGroupInputNumberArgs } from '../form-groups/input-number/form-group-input-number.args';
+
+export const inputNumberContent: FormGroupInputNumberArgs = {
+  id: 'mijn-id',
+  label: 'input number',
+  count: 99,
+  required: false,
+  disabled: false,
+  readonly: false,
+  minusButtonInactive: false,
+  plusButtonInactive: false,
+  min: 0,
+  max: 99
+};