+++ packages/sources/src/components/form/form.stories.ts
@@ -7,14 +7,16 @@ import { formArgsMapper, formArgTypes } from './form.args';
 import { checkboxesContent } from './content/checkboxes.content';
 import { confirmContent } from './content/confirm.content';
 import { inputContent } from './content/input.content';
+import { inputNumberContent } from './content/input-number.content';
 import { radiosContent } from './content/radios.content';
 import { selectContent } from './content/select.content';
 import { staticContent } from './content/static.content';
 import { textareaContent } from './content/textarea.content';
-import { Form, FormGroupCheckboxes, FormGroupConfirm, FormGroupInput, FormGroupInputDate, FormGroupRadios, FormGroupStatic, FormGroupSelect, FormGroupTextarea } from './form.models';
+import { Form, FormGroupCheckboxes, FormGroupConfirm, FormGroupInput, FormGroupInputDate, FormGroupInputNumber, FormGroupRadios, FormGroupStatic, FormGroupSelect, FormGroupTextarea } from './form.models';
 import { formGroupCheckboxesArgsMapper, formGroupCheckboxesArgTypes } from './form-groups/checkboxes/form-group-checkboxes.args';
 import { formGroupConfirmArgsMapper, formGroupConfirmArgTypes } from './form-groups/confirm/form-group-confirm.args';
 import { formGroupInputArgsMapper, formGroupInputArgTypes } from './form-groups/input/form-group-input.args';
+import { formGroupInputNumberArgsMapper, formGroupInputNumberArgTypes } from './form-groups/input-number/form-group-input-number.args';
 import { formGroupRadiosArgsMapper, formGroupRadiosArgTypes } from './form-groups/radios/form-group-radios.args';
 import { formGroupSelectArgsMapper, formGroupSelectArgTypes } from './form-groups/select/form-group-select.args';
 import { formGroupStaticArgsMapper, formGroupStaticArgTypes } from './form-groups/static/form-group-static.args';
@@ -27,6 +29,7 @@ export interface FormParameters<TemplateFnReturnType> {
   formGroupCheckboxesTemplate: (formGroupCheckboxes: FormGroupCheckboxes<TemplateFnReturnType>) => TemplateFnReturnType;
   formGroupConfirmTemplate: (formGroupConfirm: FormGroupConfirm<TemplateFnReturnType>) => TemplateFnReturnType;
   formGroupInputTemplate: (formGroupInput: FormGroupInput<TemplateFnReturnType> | FormGroupInputDate<TemplateFnReturnType>) => TemplateFnReturnType;
+  formGroupInputNumberTemplate: (formGroupInputNumber: FormGroupInputNumber<TemplateFnReturnType>) => TemplateFnReturnType;
   formGroupRadiosTemplate: (formGroupRadios: FormGroupRadios<TemplateFnReturnType>) => TemplateFnReturnType;
   formGroupSelectTemplate: (formGroupSelect: FormGroupSelect<TemplateFnReturnType>) => TemplateFnReturnType;
   formGroupStaticTemplate: (formGroupStatic: FormGroupStatic<TemplateFnReturnType>) => TemplateFnReturnType;
@@ -45,6 +48,7 @@ export function storiesOfForm<TemplateFnReturnType>(
     formGroupCheckboxesTemplate,
     formGroupConfirmTemplate,
     formGroupInputTemplate,
+    formGroupInputNumberTemplate,
     formGroupRadiosTemplate,
     formGroupSelectTemplate,
     formGroupStaticTemplate,
@@ -95,6 +99,18 @@ export function storiesOfForm<TemplateFnReturnType>(
     })
     .addDecorator(formGroupDecorator)
     .add('input', bindTemplate(formGroupInputArgsMapper, formGroupInputTemplate)
+  );
+
+  storiesOf('Form/groups/input-number', mainModule)
+    .addParameters({
+      argTypes: formGroupInputNumberArgTypes,
+      args: inputNumberContent,
+      docs: {
+        page: readme
+      }
+    })
+    .addDecorator(formGroupDecorator)
+    .add('input number', bindTemplate(formGroupInputNumberArgsMapper, formGroupInputNumberTemplate)
   );
 
   storiesOf('Form/groups/radios', mainModule)